TensorRT refactor - bust the monolith
